net: framework: Remove framework/net/pack
