--- src/framework/net/pack/Mybuild
@@ -1,7 +1,0 @@
-package embox.framework.net
-
-module pack {
-	source "core.c"
-
-	depends embox.util.Array
-}
