+++ src/net/net_pack.c
@@ -7,9 +7,9 @@
  * @author Ilia Vaprol
  */
 
-#include <framework/net/pack/api.h>
 #include <stddef.h>
 #include <util/array.h>
+#include "embox/net/pack.h"
 
 ARRAY_SPREAD_DEF(const struct net_pack, __net_pack_registry);
