+++ src/net/socket/raw_sock.c
@@ -8,8 +8,11 @@
  */
 
 #include <assert.h>
-#include <embox/net/sock.h>
+#include <stddef.h>
+#include <string.h>
+#include <sys/socket.h>
 #include <errno.h>
+
 #include <net/l3/ipv4/ip.h>
 #include <net/l3/icmpv4.h>
 #include <net/skbuff.h>
@@ -17,10 +20,11 @@
 #include <net/socket/inet_sock.h>
 #include <net/socket/raw.h>
 #include <netinet/in.h>
-#include <stddef.h>
-#include <string.h>
-#include <sys/socket.h>
+
 #include <util/dlist.h>
+
+#include <embox/net/sock.h>
+#include <embox/net/pack.h>
 
 static const struct sock_proto_ops raw_sock_ops_struct;
 const struct sock_proto_ops *const raw_sock_ops
